Renaming webpack/config to webpack/app
